Remove simple "indented" declaration styling.
This rewrites the logic used to format/indent Swift declarations. The
new logic simply breaks each parameter onto its own line (except when
there is only a single parameter).

For context, the old way of formatting Swift declarations was different for
Swift-only vs Swift+ObjC symbols was as follows:
  * Swift-only was handled very minimally with the "indented" styling
  where the full declaration wrapped vertically, with some visual space
  for every line following the first
  * The Swift code for Swift+ObjC symbols was formatted in a way that
  vaguely resembles the formatting for ObjC symbols where the
  identifiers (function name + external param names) broken apart and
  visually aligned in a vertical manner

This change aims to be an improvement in readability for Swift-only
symbol declarations while also providing consistency in the way that
Swift declarations are formatted for Swift+ObjC symbols.

Any further improvement to this indentation formatting would likely
require some tool that works with an AST (like SwiftFormat/SwiftSyntax)
as opposed to just an array of tokens for syntax highlighting.