Prevent crash when index response is empty.
Currently, if the request for the navigator data returns an empty
response like so:

```json
{ "interfaceLanguages": {} }
```

Then the sidebar shows the the "Fetching..." message indefinitely, even
though the data has already been fetched. This happens because a crash
in the code where it is assumed that there will always be Swift data to
fall back on, which is not always the case.

Now, in this scenario, the code won't crash, and the message will say
"No data available." in the sidebar to better indicate the situation.