Calculate intrinsic image width w/o `naturalWidth`.
Since `naturalWidth` output is different across browsers due to a Chrome
optimization based on 2x being in the filename, this is an alternate way
of calculating the actual image width on the client that requires
loading the image again behind the scenes. It also results in a quick
flash in instances where the image needs to be scaled down.