Fix bug where nav toggle can reset theme settings.
There is code in the function to lock the scroll which adds inline
styles to the body element. These styles are removed by clearing out all
the inline styles when the unlock function is called—but also any inline
styles are removed which interferes with the theme settings variables—by
being more explicit about which inline styles are removed, this bug can
be avoided.  Now, only those styles needed for locking/unlocking scroll
are removed when that function is called.