Ensure user always scrolls to right location of URL fragment.
There were 2 main changes needed to make this work right everywhere:

1. Update the `onPageLoadScrollToFragment` mixin to also scroll in the
`updated` hook as well as the existing `mounted` one since that is the
appropriate hook for the situation where a user clicks a link to an
anchor on another page within the same view.

2. Eagerly load all images so that the scrolling code can wait until
they've all been loaded before scrolling since the dynamic heights of
the images will impact the location that needs to be scrolled to.

As a future optimization, we should also try to see if we can use lazy
loading as the default for images and only eagerly load them when a URL
with a fragment is routed to.