Fix Safari bug with language toggle width. (#517)
In the latest Safari Technology Preview, the language toggle is not
being re-sized as expected when toggling between languages. This seems
to be an issue with the combination of this new browser and Vue where
the HTML <option> element is not being properly updated like it should
when the toggle is used. Adding a `key` attribute fixes this, because it
forces Vue to not attempt to re-use the same element for unique language
choices.

With the option element updating properly, the appropriate size for the
dropdown can be calculated, and the visual issue is resolved.

Resolves: rdar://104194977