Use `AbortController` to limit excessive fetches.
If a network is slow and a user is quickly scrolling through the list of
symbols, the pending fetch requests can pile up. By using an abort
controller, we can cancel any requests still pending for symbols that
aren't being focused on at the moment to better optimize resources on
the currently selected symbol and the one immediately following it.