Remove min-height rule from `CodeListing`.
This causes issues when code listings are used in `@Row`/`@Col` DocC
directives, because it results in them expanding beyond their normal
size and potentially obscures content below the row.
It doesn't seem that the `min-height` rule is still needed with the
recent updates to the styling of code listings whereas it was previously
used to fix an issue with overflow scrolling in tutorial code.