utils: perform portable path sanitisation of URLs
Some file systems have restrictions on character sets which are valid
file name characters.  Add a filter for the Windows file system
character set restrictions.  We replace them with `_` to match the
behaviour in the DocC bundle generation after apple/swift-docc#668.

Conditionally enable the new portable paths based upon a setting in
`theme-settings.json` to provide a means for compatibility.