Fix broken import for non-custom syntax langs.
This is a regression mistakenly introduced with [#724][0] that prevents
non-custom syntax languages (everything except swift/markdown) from
being highlighted as expected due to an `import` call that is throwing
an exception.

To fix it, we can simply append the `.js` file extension to ensure that
the import continues to work as expected.

[0]: https://github.com/apple/swift-docc-render/pull/724