Remove ability to override icons in "Topics" section.
Due to limitations with remote SVG assets, there can be an issue where
the color of a custom icon may be present in multiple locations using
the same color, which results in a less than ideal experience.

Current known workarounds for this issue would necessitate server side
headers for cross origin resource sharing, which DocC-Render can't
control since it is a client-side only tool—with that in mind, it makes
more sense to remove the ability for overriding the icon here and
leaving it as something to customize in the hero element only for now.