Add `Pager`.
Introduces a new "pager" component, which can be used to group pages of
content with only one page being visible at any given time. Users can
toggle the page of content currently being shown.

As an experiment, this pager component is being used on the grid-style
`@Links` so that only a set number of links are shown at one time.