Use `includedArchiveIdentifiers` to inactivate certain links.
When multiple targets that reference each other are compiled together
using the experimental DocC support, this array will indicate which
targets are available for linking purposes.

This allows the renderer to inactivate links in the situation where a
single target is being presented on its own and therefore cannot link to
the other target.