Prefer modules when multiple nav trees provided in index.
This addresses a rare edge case where the Index JSON contains multiple
top-level root nodes for distinct trees of navigation data. While that
shouldn't normally happen in practice, we would like to prefer the tree
with a containing module if available instead of always using the first
tree provided in the array.