[6.2] Add HTML data attributes to background and hero elements (#942)
* Add HTML data attributes

We want to add HTML data attributes [1] to background and hero elements
to be able to recognize them. We are only interested in hero that have
an enhance background, so in the case that it doesn't have it, the data
attribute won't be rendered.

[1] https://developer.mozilla.org/en-US/docs/Web/HTML/How_to/Use_data_attributes

* Fix tutorial background

Add a default background-color for tutorials. Also, we want to remove
any padding or margin of tutorial background for IDEs.