Start to use CSS logical properties for docs pages. (#954)
This transition to logical properties and values (as opposed to
physical/dimensional ones) will allow for UI elements and associated
spacing to better accomodate languages with a right-to-left writing
direction.

Although DocC doesn't yet support right-to-left localized content, this
change will better future-proof the renderer if it does, allowing it to
do things like flip the orientation of the sidebar and strings of text
for languages that read from right to left instead of left to right.

https://developer.mozilla.org/en-US/docs/Web/CSS/CSS_logical_properties_and_values

Resolves: rdar://138952234