Normalize relative path
For the normalizePath function to work, the path needs to be relative.
This means that we need to normalize the relative path first, to make
sure that the normalizePath function reads it as a relative path.

For this, we should use the function isAbsoluteUrl to see if the url is
absolute or not.