Flatten modules
In some cases, our navigator can have nested modules within the top
level modules. We want to extract all of them to be able to identify
which module we want to use.

This change introduces a recursively function that flatten a nested
module structure into a single array. So we can identify which it's the
top level module.