Replace window.location.href with window.location.pathname (#983)
In the use cases of navigators with multiple top-level root, we want to
make sure that queries don't interfere with the election of the
top-level root node when extracting the root node.

Before we were reading the window.location.href, which includes queries.
Now, by using the window.location.pathname, we ignore the queries, which
makes possible to compare with the node.path value to find which it's
the correct top-level root.

Example: URL: 'http://localhost:8080/documentation/foo?language=objc'
Navigator path: '/documentation/foo'

By using `window.location.pathname` we ignore `?language=objc` and we
can compare `/documentation/foo` with `/documentation/foo`