Normalize Sample Code URLs
We want to normalize Sample Code URLs, by adding the baseUrl to them if
the path is relative and the baseUrl is provided. If the path is
absolute or the baseUrl is not provided, it does not prefixes the URL.

Because the component CallToActionButton.vue is not only being used for
Sample Code links, we created a new prop called linksToAsset that will
normalize the relative URLs, in the cases that they are assets, like
Sample Code.