Identify the root module of the navigator out of nested modules (#981)
* Rename langData as modules

It's confusing to use `landData` and `data` when we are referring to
`modules` in the navigator instead. This change replace `landData` and
`data` with `modules`. It also replace `node` with `module`. This change
will help us to identify the modules that we need to flat in future
commits.

* Flatten modules

In some cases, our navigator can have nested modules within the top
level modules. We want to extract all of them to be able to identify
which module we want to use.

This change introduces a recursively function that flatten a nested
module structure into a single array. So we can identify which it's the
top level module.

* Rename extractRootNode with extractRootModule

To be consistent with previous changes, we want to rename this function.

* Rename rootModule

Now that we aren't always extracting the top level modules, but the root
module for the navigator. We want to replace the name of topLevelModule
with rootModule.

* Add unit test to the flattenModules function